requirements.txt
streamlit
pandas
langchain
langchain-openai
langchain-experimental
python-dotenv
tabulate
